refactor(hardhat): use uint16 for slippage

+++ packages/hardhat/contracts/SifiV1Router01.sol
@@ -39,7 +39,7 @@ contract SifiV1Router01 is Ownable {
     uint256 amountOut,
     address[] memory path,
     address to,
-    uint256 slippage,
+    uint16 slippage,
     uint256 deadline
   ) public payable returns (uint256[] memory amounts) {
     path[0] = address(weth);
@@ -71,7 +71,7 @@ contract SifiV1Router01 is Ownable {
     uint256 amountOut,
     address[] memory path,
     address payable to,
-    uint256 slippage,
+    uint16 slippage,
     uint256 deadline
   ) public returns (uint256[] memory amounts) {
     spender.transferFrom({token: path[0], from: msg.sender, to: address(this), amount: amountIn});
@@ -116,7 +116,7 @@ contract SifiV1Router01 is Ownable {
     uint256 amountOut,
     address[] memory path,
     address to,
-    uint256 slippage,
+    uint16 slippage,
     uint256 deadline
   ) public returns (uint256[] memory amounts) {
     spender.transferFrom({token: path[0], from: msg.sender, to: address(this), amount: amountIn});
